requirements.txt
# Core Agent Framework
agno>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Vector Similarity Search (RAG)
faiss-cpu>=1.7.4
# For GPU support, use: faiss-gpu>=1.7.4

# Embeddings & NLP
sentence-transformers>=2.2.0
torch>=2.0.0

# Evaluation & Metrics
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework
streamlit>=1.28.0

# Utilities
duckduckgo-search>=3.9.0
pypdf>=3.0.0
pyyaml>=6.0.0

# Optional (for advanced features)
lancedb>=0.3.0
tantivy>=0.20.0
yfinance>=0.2.0

# Progress bars
tqdm>=4.66.0